+++ packages/core/src/table/types.ts
@@ -33,6 +33,9 @@ export type ReactTanstackTableUiProps<T> = {
   tableRef?: React.RefObject<RttuiRef | undefined>;
   autoCrushNumCols?: number;
   shouldUpdate?: ShouldUpdate;
+  debug?: {
+    measureInstance?: boolean;
+  };
 };
 
 export type UiProps = Omit<