+++ chadcn-registry/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rttui/chadcn-registry
 
+## 0.0.2
+
+### Patch Changes
+
+- add ability to click rows and add a measure debug option
+
 ## 0.0.1
 
 ### Patch Changes